Add option to shorten multi-party chat names.
This commit implements the proposal laid out in issue #96.

By default, multi-party direct message chats are directly mapped to an
IRC channel name that matches this scheme:

  #mpdm-mynick--nick1--nick2--nick3

This is pretty unwieldly.  To try and make these channel names a bit
more manageable, this changeset adds a new preference:

  short-group-chat-names

If this is supplied with the login token, these chat names are instead
mapped to this scheme:

  &nick1-nick2-nick3

This uses the server local chat prefix to differentiate between normal
and group chats, removes the current user nick, and then takes out
some extra hyphens to compress the whole thing down.

This is made optional through a preference so as not to disrupt the
current UX.